+++ backend/src/main/kotlin/com/nicolaspetras/splfantasy/service/scorer/calculator/PointsCalculator.kt
@@ -1,10 +1,11 @@
 package com.nicolaspetras.splfantasy.service.scorer.calculator
 
 import com.nicolaspetras.splfantasy.model.stat.collection.SplPlayerStats
+import java.math.BigDecimal
 
 interface PointsCalculator {
     /**
      * Calculate the points
      */
-    fun calculatePoints(playerStats: SplPlayerStats): Double
+    fun calculatePoints(playerStats: SplPlayerStats): BigDecimal
 }
